保存・更新機能の実装

+++ src/database/migrations/2026_04_05_000240_user_movie_statuses.php
@@ -22,7 +22,8 @@ class UserMovieStatuses extends Migration
             
             $table->string('status'); 
             $table->text('user_comment')->nullable();
-            $table->integer('rating')->nullable();
+            // $table->integer('rating')->nullable();
+            $table->tinyInteger('rating')->nullable();
             $table->timestamps();
 
             // これで型が完全に一致するので通るはずです
